+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Add appearance mode button in app header
 - Add description section for import/export feature
 - Export collections for Pinner or Save Pinned Tabs
 - Import collections from Pinner or Save Pinned Tabs